server/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
requests>=2.32.0
numpy>=2.2.0,<3.0.0
pillow>=11.0.0
opencv-python>=4.11.0
transformers==4.56.0
accelerate>=1.10.0
easyocr>=1.7.2
manga-ocr>=0.1.14